requirements.txt
discord.py==0.16.12
PyNaCl==1.3.0
pandas
dnspython==1.16.0
async-timeout==3.0.1
async==0.6.2